pyproject.toml
[tool.poetry]
name = "adviceai"
version = "0.1.0"
description = ""
authors = ["dinmukhammed <arystambekdimash005@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.352"
python-dotenv = "^1.0.0"
openai = "0.28"
aiogram = "^3.2.0"
googletrans = "4.0.0-rc1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
